error on unknown names
Summary:
This diff makes us start erroring on unknown names on the static compiler. The mechanism is simple: When visiting a name, if we can't resolve a type for it, we error. The hard work is patching up the rest of static python to ensure we have a complete understanding of all names.

One detail is that we sometimes need to visit names at potential declaration points to handle reassignments, etc. (think `a = 1; a = "str"`). To avoid erroring in that situation, we keep track of whether we're visiting an assignment target.

There is one current issue in type binding: We don't properly resolve names from the outer scope in comprehensions and nested functions. This is due to some unsoundness where inner scopes can modify outer scopes, so we treat those as dynamic. We examine scopes to see if the names are present to avoid throwing spurious type errors, even if we can't trust the annotations from outer scopes.

Reviewed By: carljm

Differential Revision: D35328977

fbshipit-source-id: 057ce9b52cd816e2a34bdd306cb8f4f77c598d75