Skip flaky test in test_xmlrpc
Summary:
Very occasionally `test_xmlrpc` can get stuck. Unlike the similar issue in `test_asyncio` this was a lot harder to track down as it seems to only occur when run in a full test run with lots of other activity and even then, infrequently.

Eventually I was able to identify the specific test I've disabled in this diff with the following script:
```
RE='test_xmlrpc \([0-9][0-9][0-9]+s, pid ([0-9]+)\)'
unset fired
while true ; do
  echo DATE: $(date)
  ./rebuild.sh --force  # My local script for reconfig + building Cinder
  make -j testcinder_jit | while read line ; do
    if [ -z "$fired" ] && [[ $line =~ $RE ]] ; then
      tmux split gdb -p ${BASH_REMATCH[1]}
      fired=1
    else
      echo $line
    fi
  done
done
```

This script popped interactive `gdb` in a `tmux` split and I was able to see which specific test was running at a time when `test_xmlrpc` was stuck.

Using `gdb` I could see we were getting stuck doing a TCP receive in the final call to `p.method()` at the end of the test. My guess is under severe conditions we hit the one second timeout in the `run_server()` function above and this somehow breaks things. I didn't care to investigate further and just disabled the test as I don't think this particular test is that useful for Cinder development.

Reviewed By: itamaro

Differential Revision: D35507607

fbshipit-source-id: d13e8af8d7e3e0f7bec1bd8a17f83d279454bae5