None should have __class__
Summary: Fixes an issue that came up in the codemod for D35353130 - we report errors for all attributes on `None` which isn't really right.  We should recognize everything that is on `object`.  But we don't currently recognize `getset_descriptors` when we reflect over types, so we don't include `__class__`.  This adds support for the get set descriptors, but we also need to specialize the `__class__` case because we rely upon it if you do `if isinstance(..., foo.__class__)` where we know the type of `foo`.

Reviewed By: wmeehan

Differential Revision: D35483196

fbshipit-source-id: 1b6607b28bcb7103332d93ef2289625c4bd31772