Unify typed descriptor with default and property implementation
Summary:
Currently our typed descriptor with default value is practically identical to properties, but we put it under a different suffix.  Ultimately that prevents us from allowing a class attribute to be overloaded with a property, which we even have a test case for.  That caused a failure in the conversion for D35425486.

This diff simply unifies the two implementations under the `fget` suffix so that we can support the overrides in the next diff.

Reviewed By: carljm

Differential Revision: D35487353

fbshipit-source-id: e24d7e7539fd64e113b5ef3cbf5aef01b67214ba