Add @dataclass to Static Python
Summary:
This diff introduces a very bare-bones `dataclass` intrinsic to `__static__`.

When resolving the decorator,
remove it from the AST so that we do not also execute it at runtime.
Then return a wrapper around the original class,
so we can delegate most methods to it.

Dataclasses require type binding and codegen for functions without AST,
so the wrapper takes over both of these during normal class traversal.
For type binding, we create a `BuiltinMethodDescriptor`,
since built-in types also have methods without AST.
Codegen is done by hand, creating a fake `FunctionScope` and `FlowGraph`
and emitting to the graph one byte code at a time.

This diff limits extra binding and codegen to `__init__` for now.

We also ban inheritance of static data classes for now,
to limit complexity in this diff.

Reviewed By: carljm

Differential Revision: D34830211

fbshipit-source-id: c70b5b471211bd61418c371a9e413ee8260e95d6