Derive Yield instructions from DeoptBase
Summary:
Most uses of `Yield*` instructions during lowering emit an exception check
immediately following the `Yield*` (the lone execption is the `YieldFrom`
that is emitted in the header of an `async for` loop, more details below).
Other HIR instructions that follow this pattern inherit from `DeoptBase`.
This diff updates the `Yield` variants to do so as well.

The one complication that arose is that `YIELD_FROM` is used to
implement `async for` loops, wherein`StopAsyncIteration` is raised to indicate
that the loop has completed. Naively using a `YieldFrom` instruction that
inherited from `DeoptBase` in such a scenario would cause us to deopt into
the interpreter whenever an `async for` loop completed. To handle this, I've introduced
a new HIR instruction, `YieldFromHandleStopAsyncIteration`, that is used in place
of `YieldFrom` in such scenarios. It has the same semantics as `YieldFrom`,
except that it detects `StopAsyncIteration` exceptions, clears them, and
instead produces the sentinel value that we use to signal that iteration has
completed.

In addition to simplifying the generated HIR, this allows us to
remove the special logic in RefcountInsertion that saves live values
into the `Yield` variants, and will make it easier to share program
point information that is needed by both deopt and the frame introspection
infrastructure.

Reviewed By: ATalaba

Differential Revision: D35441775

fbshipit-source-id: d907da1efa34fb206a6468976f6ce8703e51f595