Allow overriding class variables with properties
Summary:
This enables overriding class variables with properties and vice-versa.  At the static level we enforce the types on the overrides, dynamically we already have the enforcement.

The biggest part of this change is cleaning up our override enforcement in general.  It now becomes a method on the base class which receives the override and allows it to do the enforcement.  This replaces a hodge-podge of enforcement decisions which were strung together over a couple of methods.  By default we require that overrides are the same type as before.

This should unblock D35425486 to hit the next error or successfully compile.

Reviewed By: carljm

Differential Revision: D35488515

fbshipit-source-id: 2861d0e8adde4f55da260794e7883668eafea69c