introduce a multiple section code allocator
Summary:
This diff adds support for a code allocator which allows Cinder to emit code into multiple sections (initially, hot and cold).

The basic idea is to allocate a large slab of memory where all code will go, and subdivide this region internally between hot and cold sections. When allocating code, we now keep a pointer to active section. We try to use huge pages, and fall back to regular mmap on failure.

The meat of the diff is the algorithm for writing code to multiple sections. We now have to deal with the jumps we wish to emit not matching asmjit's internal buffers in deopt cases. To support multiple sections, we implement a variant of asmjit's flattening algorithm when calculating offsets of different sections: We start maintaining a mapping of where each section start is before we've written the JITted function in memory, lay things out so that hot code always comes before cold and adjust offsets relatively, memset and bump the pointers afterwards.

We also change all label offsets to be relative to use `labelOffsetFromBase`. This ensures offsets are correct now that we use multiple sections (beforehand, we would only have one text section, so its offset would always be 0).

The usage of multiple code sections is gated behind flags and disabled by default. The idea would be to add these to the consumer application's env flags when starting the interpreter.

If the slabs run out of memory, we revert to asmjit's regular code allocation.

As for the naming choice: Even though the initial code  sections will be hot and cold, prior art in HHVM has shown that we might need to support more sections that we want to support. Towards this end, the naming and algos have been designed to be extensible should we want more sections in the future.

Reviewed By: swtaarrs

Differential Revision: D35015003

fbshipit-source-id: 38d2c21bacd6d1dc271ae1791a68c57200c57613