support multiple sections in disassembly
Summary:
Now that we can potentially emit code to multiple sections, we want to ensure that disassembly handles these cases correctly.

The previous annotations.cpp file assumed that there was a single text section, but most code worked as is. The strategy taken is simply to support reading multiple text sections by passing a CodeSection into asmjit, and filtering annotations by the ones that correspond to the section we're currently looking at. Since this relies on `labelOffsetFromBase`, we need to ensure that the code was already allocated, so add a JIT check to that effect.

We had previously discussed potentially also emitting all hot code + all cold code for a given function as an alternate design choice. I still went ahead with this as the disassembly reads reasonably intuitively to me.

Reviewed By: swtaarrs

Differential Revision: D35083574

fbshipit-source-id: 8e53c436d0736c0ef4a56453d796d8884c063607