modify printer to show code sections
Summary:
Since most code sections will be hot, let's default to only printing cold sections for now. We can revisit if the split becomes more even, eventually.

As part of this change, I realized that it's not a great idea to have a distinction between the text section names used by asmjit and what we report - the custom "hot"/"cold" identifiers can be misleading since the names used internally are ".text" and ".coldtext", so I chose to not worry about mapping those back to hot/cold and instead just print out the internally used section name.

Reviewed By: mpage

Differential Revision: D35186027

fbshipit-source-id: 2e28611937258649943076363d1d9cb16941faa2