Bring over compiler module
Summary:
This is the equivalent of D23434419 for cinder 3.8. It brings over the latest state
of the `compiler` module, but only the base compiler that compiles to match
upstream; thus the `strict/`, `static/`, and `readonly/` directories are omitted, as are
various `Cinder*` classes in `pycodegen.py`, `pyassem.py`, and `symbols.py`, and
the `opcode_cinder.py` module. These Cinder-specific features all need to be
brought over along with matching changes in the runtime, so they'll come later.
(All these cinder-specific features were added in 3.8, so there will be later porting
tasks for them.)

Tests are brought over too (except for `test_strict`, `test_static`, and
`test_readonly`). Only the ones that pass are enabled; the others are commented
out for now in `Lib/test/test_compiler/__init__.py`. Next step is to
bring the compiler up to date for 3.10 and re-enable the remaining tests.

I added a brief note at the top of `test_compiler/README.md` to clarify the
current maintenance of this compiler codebase.

Reviewed By: itamaro

Differential Revision: D35632243

fbshipit-source-id: 160b19390ca450ae5fc9483c93f24e1a47ccafc6