Move load method scratch to register in LIR and HIR
Summary:
This moves the load method scratch register out of being a specific location on the stack, and instead uses a plain old register.  Because this needs to be used when we deopt it means that we actually need to make the register available at HIR in addition to being available in LIR.  To accomplish this a new `GetLoadMethodInstance` HIR instruction is added, which compiles down to a LIR instruction `Load2ndCallResult`.

This gets a little bit complicated because we need to ensure that no interleaving instructions are going to get added (e.g. decref's) between the two HIR instructions.  To ensure against this we make sure that the HIR instruction captures the inputs to the LoadMethod so that the lifetimes of the variables are extended and we won't insert the decref.  We then have a validation pass after all the HIR passes to make suer this is true.

The upsides to this diff are that we can now support `LOAD_METHOD`/`CALL_METHOD` everywhere - if it's nested it's fine, we're just using different registers.  If there's an interleaving yield things are also fine.  The downside is we lose some specialization on the invoke path where we had the different kinds, and need to go through one level of indirection.

Reviewed By: tiansihu

Differential Revision: D35376939

fbshipit-source-id: b9efbbdef9c3b0f988de09cec7a890eb9c102f85