Fix @dataclass() resolution
Summary:
We were falling back to dynamic behavior when using `dataclass()` with a call.

Fix this by binding the result type of calling `dataclass()` to itself,
to mimic how `dataclasses.dataclass` returns a curried function when called with params.

Since we no longer need `Dataclass` to be aware of the `ClassDef`,
remove it from the call to `resolve_decorate_class`.

Reviewed By: carljm

Differential Revision: D35687515

fbshipit-source-id: 9a80a68c4717d029bf64f553dada0fc64e5d0959