Add caching to special thunks
Summary:
This adds a cache for our special thunks used for `fget` and `fset` methods.  There's now a `vt_specials` cache on the vtable, and our lookups are now always based on the type.  Previously we'd pass a dictionary in, sometimes the `vt_original` dictionary.  If we continued to do that we'd cache values when we're looking up an original function.

Luckily this is easy to avoid because once `vt_original` exists it is fully populated, so we can always just lookup directly in it.

We've discussed this in previous contexts, but the primary motivation for this is the next diff.  That'll cleanup some handling around "original" vs "current" values where we get confused and pick the wrong method to refer to.  It also will unify our slot setting so that the patch case is identical to the init case to avoid a proliferation of states.  In order to support that it's important to be able to use object identity comparisons against the original, which not caching the values thwarts.

Reviewed By: sinancepel

Differential Revision: D35756722

fbshipit-source-id: f9faa3b856554c7d6b6cfac503b4339c7f2db272