Implement frozen dataclasses
Summary:
Frozen dataclasses ban field mutation and deletion as best they can in Python,
by generating `__setattr__` and `__delattr__` methods that raise an exception.

We do this too, and we go one step further by raising a `TypedSyntaxError`
if users attempt to mutate or delete the attribute directly.

No matter whether the dataclass is frozen, we allow `object.__setattr__`
and `object.__delattr__` to modify the instance, with appropriate type checks.
Unlike dynamic Python, we do not have to call `object.__setattr__` in `__init__`,
since we emit `STORE_FIELD` instructions directly during codegen.

One limitation of compiling the dataclass statically is that we do not have
direct access to the class object or to `dataclasses.FrozenInstanceError`.

For the class object, we currently load it from global scope using its name.
In the future, this can be accommodated by a single exact `CAST`.

We can import `FrozenInstanceError` once per class, save it as a hidden attribute,
and load it off the class when we call `__setattr__` or `__delattr__`.

Reviewed By: DinoV

Differential Revision: D35650310

fbshipit-source-id: 07e2f0f7d1f0f185edae088c300d9e1681d3ca25