Implement dataclass __repr__
Summary:
In dynamic Python, dataclasses create their repr function by executing a string
and calling (a copied implementation of) `recursive_repr` to wrap it,
thereby protecting it from recursion.

Implement the static repr function using similar bytecode to what the string generates,
but optimizing it to take advantage of `LOAD_TYPE` and slots. We also wrap the inner
function with `reprlib.recursive_repr` during class construction.

This diff also fixes a bug that set a `Dataclass` instance to the wrapped class's instance,
when we should be creating a new instance for the `Dataclass` instead.

Reviewed By: sinancepel

Differential Revision: D35683341

fbshipit-source-id: 8d72f512f884fecdf2954bdce4f89ac2f5e3e1ea