Implement dataclass __hash__
Summary:
Dataclass hashing follows a complicated heuristic involving the `unsafe_hash`, `eq`, and `frozen` parameters
and whether or not the class has an explicitly defined `__hash__` method.

If the dataclass has an explicit `__hash__`, we cannot generate one, so error if `unsafe_hash=True`.

Otherwise, set `__hash__ = None` if `eq=True` and `frozen=False`, i.e. instances are mutable.

Generate a `__hash__` that delegates to the tuple of fields if `eq=True` and `frozen=True` or if `unsafe_hash=True`.

Reviewed By: sinancepel

Differential Revision: D35854346

fbshipit-source-id: 52e4e24c8b546c2bbe34494abfc309d92a9dc586