Introduce a fixed point for the preloader
Summary:
The current multithreaded compilation architecture bails out on recursively compiling functions if the recursively compiled one isn't preloaded. This is causing us to not compile a bunch of functions pre-fork.

The biggest reason for this is that with lazy imports, we sometimes end up evaluating more imports during the preloading of a unit, and discover more units to preload, but due to the way things are set up, we're too late and bail on preload/compiling them.

This diff changes things up so that we do a fixpoint computation of pre-loading - if we see a unit that needs to be preloaded but got loaded too late, simply repeat the preload/compilation procedure until we don't get any more units to compile.

Reviewed By: carljm

Differential Revision: D35682629

fbshipit-source-id: 0c82bea9f0e90a0bb67281560727cfb23c9d2e51