Async patches also shouldn't have self forcibly passed
Summary:
The previous versions of this updated our handling for normal functions so that we no longer pass `self` to a non-descriptor.  This follows on and updates async functions as well.

We get a similar descriptor and function handler, and then `thunk_vectorcall` needs updates as well.  In the class method case we no longer strip off the argument as it'll be handled by `type_vtable_coroutine` if necessary.  For the other call we no longer re-use `type_vtable_coroutine` because we don't want to preserve the original signature for the invoke function as it could be to something like a static function, and stripping off the 1st argument would be wrong.

Reviewed By: sinancepel, wmeehan

Differential Revision: D35794695

fbshipit-source-id: 2d31434b8d081071a75fa9d456ca40c11fd9ab12