Sync enum.py with upstream
Summary:
Enums are missing a number of patches from the cinder/3.8 branch, most of which affect inheritance.

For instance, enums must be created with mixins first, then an optional data type, and finally the base enum type.

When creating an enum, we need to make sure there is at most one data type,
and that none of the base enums define any members.

This diff copy-pastes upstream `enum.py` and `test_enum.py` to get those changes.

Reviewed By: sinancepel

Differential Revision: D35908124

fbshipit-source-id: 0ad6710e2ae0badc430eb1189f5fdc7dd46027f9