Support static awaits for co-routine objects
Summary:
Removes some of the dynamic dispatch when we're awaiting on a statically invoked co-routine.  We can currently only do this with an `INVOKE_FUNCTION` because our wrapper for `INVOKE_METHOD` won't have the correct C layout, but hopefully in the future we can make that look like a co-routine.

It's not a *huge* win and may be less significant in prod as async functions are likely to do more work there.

Reviewed By: mpage, carljm

Differential Revision: D35194272

fbshipit-source-id: 4ce3ad9225cbdbe558edbc08be24fb3e5bf8f6b4