propagate refined field information when merging branches
Summary: One interesting detail here: The refined fields don't work off the bat with the join mechanism of types only getting coarser as you do more iterations, since the absence of a refined field in the entry indicates already indicates a coarses type when merging. To accomplish this, the `changed()` logic with refined fields needs to be a bit more involved than simply structural equality checking.

Reviewed By: wmeehan

Differential Revision: D35615507

fbshipit-source-id: ce3796b64612014e346cb37be43462bc000991ad