Clear awaiter for gathered awaitables on first exception
Summary:
Clear the awaiter for gathered awaitables when one of the gathered
awaitables completes with an exception and the gather is configured
to propagate exceptions. When gather is configured to propagate
exceptions (the default) the first raised exception will be immediately
propagated into the task that awaits the gather, while the other
gathered awaitables will continue running. If the gather was awaited,
the still-running awaitables will have a borrowed reference to the
coroutine that is awaiting the gather set as their awaiter. That coroutine
may complete and be freed once the exception is propagated, leaving
a pointer to an already-freed coroutine as the awaiter of
the awaitables that are still running.

Reviewed By: vladima

Differential Revision: D35342708

fbshipit-source-id: bbaba2f6425d89500ee99ed6d772604374988fa5