Allow non-constants in static Enum
Summary:
Enum values are often defined in terms of other enum values,
to ensure that they stay in sync despite being defined in different places.

This means that we need to allow for the RHS of enum definitions to be something other than an `ast.Constant`.
Allow arbitrary code execution on the RHS, as long as it binds to the correct type for `IntEnum` and `StrEnum`.

Since `Int64Enum` requires statically-known `int64_t` values, we cannot relax the requirement here.
However, we still fix a bug that was preventing annotated assignments.

Reviewed By: sinancepel

Differential Revision: D35789963

fbshipit-source-id: 4777514600a52c320d369b15e2ffc145e4fb59fe