Track code locations for inlined functions
Summary:
This diff adds support for tracking the location of all activations (inlined &
non-inlined) at any point in the generated code. This is used to provide
correct bytecode offsets when frames are materialized or when stacks are walked
using one of the internal APIs.

The high-level approach is the same as pre-inliner. For each unit on the stack,
retrieve its saved return address, which identifies the current point in the
generated code, and its `jit::CodeRuntime`. Next, retrieve location information
for the address from debug data stored in the `jit::CodeRuntime`.

There were a few major changes that were required to make this work:

First, we now store location information in a compressed form in a
`jit::DebugInfo` object in the `jit::CodeRuntime`. The `jit::DebugInfo` class
provides an API to retrieve the location information for the "call stack" at an
address in the generated code.

Second, a new kind of shadow frame, `JITShadowFrame`, was introduced for
non-generator JIT-compiled functions. This is needed so that we can recover the
`jit::CodeRuntime` from any shadow frame in the unit once the frames in the
unit have been materialized. When frames are materialized, the data field of
the shadow frame is overwritten to point at the `PyFrameObject`, leaving us no
way to retrieve the `jit::CodeRuntime` or `jit::RuntimeFrameState` that was
previously stored there. A `JITShadowFrame` adds an extra field that contains
the original value of the `data` field when the frame is materialized.

Although it adds some temporary inconsistency, `JITShadowFrame`s are only used
for non-generator functions to avoid bloating non-JIT generators. Upcoming changes
to generators will keep shadow frames in the spill data. That provides a natural
point to unify on `JITShadowFrame`.

It's worth noting that this doesn't impact our ability to recover location
information for JIT generators: we can always recover the `jit::CodeRuntime`
from a generator shadow frame. We do not support inlining in generators and
their shadow frames are embedded inside of generator objects. Given a generator
shadow frame, we can recover the generator, and from there get access to its
`GetDataFooter`, which contains a pointer to the `jit::CodeRuntime`.

Finally, our materialization and stack walking code needed to be refactored to
operate on units. Two new internal APIs, `walkShadowStack` and
`walkAsyncShadowStack`, were added to simplify this.

Reviewed By: swtaarrs

Differential Revision: D35657165

fbshipit-source-id: 5b4af01d556d53b61c5c5421e71bd63d1488b138