Don't allow boxing values of type (TPrimitive | TNullptr)
Summary:
This would already have produced weird results at runtime, due to TNullptr being allowed to represent pointers to uninitialized memory. This was likely only written this way to preserve the exact semantics of the `src_type.couldBe()` calls that were there prior.
With D35736940 the runtime type of the src register will be checked, which would produce an error if any of these conditions were hit.
This doesn't appear to actually be in use anywhere.

Reviewed By: swtaarrs

Differential Revision: D35989807

fbshipit-source-id: 627d1b0419881da1272de390dcf0128da99a5291