Fix LIR helper translation of JITRT_GetI32_FromArray
Summary: It was still assuming the elements of 32-bit arrays were 64-bit. This required updating a couple of tests that assumed that, so I've also added the 64-bit version to this table to make the tests happy. The rest of the tests were previously updated when this helper was fixed in D35760657 (https://github.com/facebookincubator/cinder/commit/d67592a6b4fe6bdc42cf6155b611d12e2c102009).

Reviewed By: mpage

Differential Revision: D36085368

fbshipit-source-id: f7c2c36301fbe880ab027649b7cf09dcf483a85b