Back out keyobject vectorcall
Summary:
This is a change we've forward ported across several Cinder versions now, and there is
little chance of us upstreaming it. It requires adding code duplication, and I can't detect
any performance benefit from it even in a very focused microbenchmark. It applies to a very
niche use case (`functools.cmp_to_key`) which is basically a Python 2->3 transition helper
that is unlikely to be used in new Python 3 code.

Rather than continuing to carry this diff forward from upgrade to upgrade, let's just remove
it.

Reviewed By: itamaro

Differential Revision: D36188834

fbshipit-source-id: 486763c3f4484ef387c4e6952a5975c3934e5e30