modify how annotations are passed to match the new expected format
Summary:
The 3.10 behavior of passing annotations is a bit different from the 3.8 version.

Instead of passing in a tuple of tuples, everything's flattened when `__future__.annotations` isn't present. When `from __future__ import annotations` is applied, the expectations is a `LOAD_CONST` with a tuple of the names. Let's change the Python compiler to work with this paradigm.

Reviewed By: carljm

Differential Revision: D35979274

fbshipit-source-id: bb4213889ce12893b942ce05046342e26aabbd68