Fall back to dynamic dataclass when base is dynamic
Summary:
If a dataclass has a dynamic base, it is possible that the base is also a dataclass.
When that happens, it is impossible to pick up the base's fields at compile time.
This means we cannot guarantee proper inheritence of fields from dynamic bases.

To ensure that we always pick up fields from our superclasses,
fall back to the dynamic `dataclass` when any base is dynamic.

Reviewed By: carljm

Differential Revision: D35981420

fbshipit-source-id: 9347bf820ac3537c084ae448ce45625fa784cf2e