Handle dataclass field assignment
Summary:
Dataclass fields currently fail type checks when assigned using `field()`.
```
dataclass
class C:
  x: str = field(metadata={"some": "data"}, default="default"}
```
In the above example, we appear to be assigning a `Field` object to a `str` field.
However, the true effect is to indicate metadata to the `dataclass` decorator,
and the actual assignment should be `x = "default"`.

Teach the static compiler how to handle `field()` assignments with optional `default` or `default_factory` arguments.

For now, we still do not attach `__dataclass_fields__` to the class
or handle the various `bool` flags that indicate whether to consider fields for certain methods.

Reviewed By: carljm

Differential Revision: D36075916

fbshipit-source-id: 20a9cfad4454e85a77a66b972f25bfd3faa7f39d