warn when dataclass has dynamic base
Summary:
When a dataclass has a dynamic base, that base might also be a dataclass.
Because we cannot determine that information at compile time,
we must fall back to dynamic behavior so that `dataclass` can pick up the information at runtime.

Warn users that they should convert all of the superclasses to Static Python,
so that the dataclass can be handled statically as well.

Reviewed By: carljm

Differential Revision: D36167223

fbshipit-source-id: 1ffb9cb384bdfe94ab188d09b1bf9a20db596a9a