hash unicode object during immortalization walk
Summary:
Currently, there are around 10% of mutation to master memory coming from unicode hashing as it is cached in the unicode object (https://fburl.com/scuba/py_mem_isolation/ojc9uooi).

This diff tries to resolve the issue by hashing all the unicode object in the heap when immortalizing the heap. This removes almost all the copy-on-writes caused. It also does all the hashing in master vs repeating in the worker process. The only trade-off is that we could be too much hashing that may not be used. Not sure if there is a way to measure that.

Reviewed By: eduardo-elizondo

Differential Revision: D35992939

fbshipit-source-id: 7c1cc839d54c9581d2d4c9a70b52ca100c5e3df0