fix ContextDecorator exception handling
Summary:
To save on the cost of raising exceptions that are never handled in Python code, setting an
exception in C will often set the type of the exception as `exc_type` and a simple string
message as `exc_val`, and traceback will be left `NULL`. Before the exception is ready for
consumption by Python exception handling code, there's a little dance we have to do:
create a traceback, call `PyErr_NormalizeException` to actually ensure the exception value
becomes an instance of the exception type, and ensure we set the traceback on the
created exception object.

Normally this dance happens in the `error:` label of the ceval loop (which is how I figured out
what it needs to do). But our specialized C code for efficient handling of ContextDecorator
calls was not doing this handling, and so if the exception was raised within the same frame
and never passed through normal frame exception handling, it would never get normalized,
and so the `__exit__` method of a `ContextDecorator` would get bad arguments.
This would show up as the real error being hidden and replaced by an error like
`TypeError: __exit__ expected 'BaseException' for argument exc_value, got 'str'`,
because the argument type checking in `__exit__` (which is a static function) would not like
the wrongly typed arguments it was getting.

Add the exception and traceback normalization code into our ContextDecorator error
handling to fix this.

Reviewed By: czardoz

Differential Revision: D36222534

fbshipit-source-id: e229a9d8cdc623788cb17e89a018adc7e78e849a