introduce an allow list regex
Summary: In larger codebases, RPC bindings such as Thrift tend to be codegenned for different languages. These bindings can have module-level effects that are hard for the strict analyzer to analyze, but are benign in practice. The codegenned files tend to conform to common naming schemes. To support using the constructs from these modules, let's enhance the allowlisting functionality of the strict analyzer to support regexes.

Reviewed By: czardoz

Differential Revision: D36285164

fbshipit-source-id: 2a57a117f6714c383d79a7326ac722edf73f73d0