Use field flags to generate methods
Summary:
Dataclass fields are allowed to opt out of being considered for each of the generated methods.

For instance, a field might set `field(init=False, default_factory=list)` if it should always be initialized to an empty list.

Fields can opt out of `__init__`, `__hash__`, `__repr__`, and the comparison functions.

Fields cannot opt out of being frozen.

`ClassVar` and `InitVar` fields should always be ignored for generated methods, with one exception:
`InitVar` fields are arguments to `__init__` and `__post_init__` unless opted out.

We must also store the base class on which each `DataclassField` is defined,
so that we have an accurate type descr for accessing fields on subclasses.

This diff also changes our internal representation of `fields` into a `dict`,
to better match dynamic behavior and accommodate future subclassing.

Reviewed By: carljm

Differential Revision: D36213817

fbshipit-source-id: c97ecdef08b010efd13cfbf7495974d842c34488